backend/pyproject.toml
[project]
name = "superin"
version = "2.1.0"
description = "Superin — SuperApp platform"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.7.0",
    "email-validator>=2.2.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.20",
    "motor==3.7.1",
    "beanie==1.29.0",
    "langgraph==1.1.4",
    "langgraph-prebuilt==1.0.8",
    "langgraph-checkpoint-mongodb==0.3.1",
    "langgraph-store-mongodb==0.2.0",
    "langchain-core==1.2.31",
    "langchain-openai==1.1.10",
    "openai==2.32.0",
    "langsmith[claude-agent-sdk]>=0.4.52",
    "python-dotenv>=1.0.0",
    "httpx>=0.28.0",
    "pytz>=2024.2",
    "boto3>=1.35.0",
    "loguru>=0.7.3",
    # Security & Sanitization
    "nh3>=0.3.4",            # HTML sanitization (Rust-based)
    "confusables>=1.2.0",    # Homoglyph detection
    # Rate Limiting (Redis-backed, multi-worker safe)
    "fastapi-limiter>=0.1.6",
    "redis[hiredis]>=5.2.0",
]

[project.optional-dependencies]
dev = [
    "pip-tools>=7.4.1",
    "pytest>=8.3.0",
    "pytest-asyncio>=0.25.0",
    "httpx>=0.28.0",
    "ruff>=0.9.0",
    "mypy>=1.14.0",
]

[project.scripts]
superin = "core.main:app"

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
filterwarnings = [
    "ignore:Please use `import python_multipart` instead\\.:PendingDeprecationWarning",
]

[tool.mypy]
python_version = "3.11"
check_untyped_defs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
warn_return_any = true
warn_unused_ignores = true

[[tool.mypy.overrides]]
module = [
    "boto3",
    "botocore.client",
    "botocore.exceptions",
    "confusables",
    "jose",
    "loguru",
    "passlib.*",
]
ignore_missing_imports = true
